Address review feedback on placement limit and shuffle tests

+++ pkg/scheduler/schedule_one_podgroup.go
@@ -664,7 +664,7 @@ func (sched *Scheduler) podGroupSchedulingPlacementAlgorithm(ctx context.Context
 	var anyResult *podGroupAlgorithmResult
 	successfulResults := make(map[*fwk.Placement]*podGroupAlgorithmResult)
 
-	numPlacementsToFind := int32(1)
+	numPlacementsToFind := 1
 	if schedFwk.HasPlacementScorePlugins() {
 		numPlacementsToFind = sched.numFeasiblePlacementsToFind(schedFwk.PercentageOfPlacementsToScore(), placements)
 	}
@@ -672,7 +672,7 @@ func (sched *Scheduler) podGroupSchedulingPlacementAlgorithm(ctx context.Context
 	// Only a subset of placements will be evaluated (the loop breaks once enough feasible ones are
 	// found), so shuffle to make that subset random rather than dictated by the PlacementGenerate
 	// plugin. Shuffle a copy to avoid mutating the slice the plugin returned.
-	if sched.shufflePlacements != nil && int(numPlacementsToFind) < len(placements) {
+	if sched.shufflePlacements != nil && numPlacementsToFind < len(placements) {
 		placements = slices.Clone(placements)
 		sched.shufflePlacements(placements)
 	}
@@ -701,7 +701,7 @@ func (sched *Scheduler) podGroupSchedulingPlacementAlgorithm(ctx context.Context
 			successfulResults[placement] = &result
 		}
 
-		if len(successfulResults) >= int(numPlacementsToFind) {
+		if len(successfulResults) >= numPlacementsToFind {
 			break
 		}
 	}
@@ -738,19 +738,19 @@ func randShufflePlacements(placements []*fwk.Placement) {
 
 // numFeasiblePlacementsToFind returns the number of feasible placements that once found, the scheduler stops
 // its search for more feasible placements.
-func (sched *Scheduler) numFeasiblePlacementsToFind(percentageOfPlacementsToScore *int32, placements []*fwk.Placement) (numPlacements int32) {
-	numAllPlacements := int32(len(placements))
+func (sched *Scheduler) numFeasiblePlacementsToFind(percentageOfPlacementsToScore *int32, placements []*fwk.Placement) (numPlacements int) {
+	numAllPlacements := len(placements)
 
 	if numAllPlacements < minFeasiblePlacementsToFind {
 		return numAllPlacements
 	}
 
 	// Use profile percentageOfPlacementsToScore if it's set. Otherwise, use global percentageOfPlacementsToScore.
-	var percentage int32
+	var percentage int
 	if percentageOfPlacementsToScore != nil {
-		percentage = *percentageOfPlacementsToScore
+		percentage = int(*percentageOfPlacementsToScore)
 	} else {
-		percentage = sched.percentageOfPlacementsToScore
+		percentage = int(sched.percentageOfPlacementsToScore)
 	}
 
 	// If neither is set or the set value is 0, linearly interpolate the value between (0, 100) and (5000, 10),
@@ -764,10 +764,10 @@ func (sched *Scheduler) numFeasiblePlacementsToFind(percentageOfPlacementsToScor
 			// which can be different than the number of nodes in the cluster.
 			numAllNodes += len(placement.Nodes)
 		}
-		percentage = max(minFeasiblePlacementsPercentageToFind, int32(100-numAllNodes*9/500))
+		percentage = max(minFeasiblePlacementsPercentageToFind, 100-numAllNodes*9/500)
 	}
 
-	numPlacements = max(minFeasiblePlacementsToFind, int32(numAllPlacements)*percentage/100)
+	numPlacements = max(minFeasiblePlacementsToFind, numAllPlacements*percentage/100)
 
 	return numPlacements
 }
